Version 4.00.00
As of Version 4.00.00 or higher, you have the option to choose between two views: PGN View (the new, added version) and PGN Trace (the legacy version). The default display is PGN View, which shows an overview of all current PGNs along with a counter, making for a smoother screen presentation. Although PGN Trace can provide more effective information by displaying PGNs in the sequence as they appear, it is sometimes more difficult to read, especially with high-frequency PGNs.