Legacy: J1939 to USB Gateway firmware for NXP LPC1754 processor
Our new design is based on the ESP32 processor, but we tried to keep the functionality identical to the NXP LPC1754 version. However, as of September 2024, we are now at a point where we need to modify the firmware to support higher communication speed with the PC, which is mandatory for another new feature, i.e., the support of timestamps for CAN data frames. These modifications apply only to the ESP32 version. Refer to the user manual for the firmware update/reload:
https://jcom1939.com/jcom1939-legacy-version/